Update description of workflow to include RawResult
Explain that the loader implementation should create RawResult
model instances and that the transforms should create Contest,
Candidate and Result records.

Separate descriptions of the modules that need to be implemented
from descriptions and usage examples of the invoke tasks that run
them.

Show code examples of registering transforms and validators.

Addresses openelections/core#49