s/invoke/openelex/
Now that we've moved from using invoke as the task runner to having
our own management command, ``openelex``, update the contributing
code guide to use ``openelex`` in the command line examples instead
of ``invoke``.  Also, remove warnings about having to be in the
package directory or the repository root since we no longer have
that limitation.